+++ foxes/input/yaml/__init__.py
@@ -5,5 +5,4 @@ from .dict import run_obj_function as run_obj_function
 from .dict import get_output_obj as get_output_obj
 
 from .yaml import foxes_yaml as foxes_yaml
-from .windio.windio import read_windio as read_windio
 from .windio.windio import foxes_windio as foxes_windio